Refactor Cartesian coordinate system: remove deprecated Cartesian module, enhance Position and Direction types with validation, and improve structure for clarity and consistency.
